+++ pstatus-report-sink-ktor/src/main/kotlin/gov/cdc/ocio/processingstatusapi/mongo/MongoRepository.kt
@@ -5,7 +5,6 @@ import com.mongodb.client.MongoClient
 import com.mongodb.client.MongoClients
 import com.mongodb.client.MongoDatabase
 import gov.cdc.ocio.processingstatusapi.persistence.Collection
-import gov.cdc.ocio.processingstatusapi.persistence.MongoCollection
 import gov.cdc.ocio.processingstatusapi.persistence.ProcessingStatusRepository
 import org.bson.BsonDocument
 import org.bson.BsonInt64
